Project Start
In Line Objects is a front-end management system inspired from
Angular.js without the restrictive nature of a heavy top-end library.
The open nature of the object allows the use of standard JavaScript or
Dom editors like jQuery and makes it elastic across other objects or
uses of the data model bound to the object. Basically this will allow
you to merge you front end HTML with a data model and push changes to
the model which will amend or interact with the html view in real time
on the client.